requirements.txt
# Core ML and Data Science
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.15.0
mlflow>=2.9.0

# Data Processing
pyspark>=3.5.0
pyspark-sql>=3.5.0
delta-spark>=3.0.0
dbt-core>=1.7.0
dbt-redshift>=1.7.0

# Streaming and Messaging
kafka-python>=2.0.2
confluent-kafka>=2.3.0
pydantic>=2.5.0

# API and Web
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
requests>=2.31.0

# Database and Storage
boto3>=1.28.38
psycopg2-binary>=2.9.7
redis==4.6.0

# Monitoring and Observability
prometheus-client==0.17.1
grafana-api==1.0.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.65.0

# Development and Testing
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0
mypy==1.5.1
